+++ src/pages/Search/SearchPage.jsx
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import { PageWrapper } from '@components/shared/UIStyles';
+import { Container, PageWrapper } from '@components/shared/UIStyles';
 import { useSearchParams } from 'react-router';
 import { useState, useEffect } from 'react';
 import SearchItem from '@components/SearchItem/SearchItem';
@@ -30,7 +30,27 @@ const SearchPage = () => {
 
   return (
     <SearchContainer>
-      <SearchItem query={query} searchedResult={searchData} />
+      <SearchHeader>
+        <HeaderSubText>
+          {searchData.length === 0 ? (
+            <>
+              <span style={{ color: '#3092FA' }}>{`{${query}}`}</span> 에 대한 검색 결과가 없습니다.
+            </>
+          ) : (
+            <>
+              <span style={{ color: '#3092FA' }}>{`{${query}}`}</span> 에 대한 공동구매들을
+              발견했어요!
+            </>
+          )}
+        </HeaderSubText>
+      </SearchHeader>
+      {searchData.length > 0 && (
+        <ResultCountText>
+          진행중인 공동구매 총&nbsp;
+          <span style={{ color: '#3092FA' }}>{searchData.length}</span>건
+        </ResultCountText>
+      )}
+      <SearchItem searchedResult={searchData} />
     </SearchContainer>
   );
 };
@@ -40,4 +60,30 @@ export default SearchPage;
 const SearchContainer = styled(PageWrapper)`
   margin-top: 45px;
   gap: 45px;
+`;
+
+const SearchHeader = styled(Container)`
+  width: 1065px;
+  padding: 5px 0;
+
+  border-top: 1px solid #000;
+  border-bottom: 1px solid #000;
+`;
+
+const HeaderSubText = styled.p`
+  color: #333;
+  ${({ theme }) => theme.fontStyles.Body6};
+  font-weight: 500;
+  line-height: 201%;
+`;
+
+const ResultCountText = styled(HeaderSubText)`
+  width: 100%;
+  height: 32.2px;
+  margin-bottom: -40px;
+
+  font-size: 14px
+  line-height:230%;
+  letter-spacing: -0.26px;
+
 `;
